feat: surface training progress in discussion feed via shared TrainingTaskCard
